fix(agent-core): provider_request logged the caller's array, not the request that was sent
Review finding on #1802 (SHOULD), and it is correct. `provider_request`
was emitted BEFORE `callProviderWithCache`, carrying
`conversationMessages`. The structured-output transport guard can add a
system instruction that array does not contain — so for a provider
without a schema parameter, the session log described a request the
model never received. `agent-session/docs/SPEC.md:281` promises this
event carries the request envelope, and a replay has to be able to
reproduce it.

The callback now reports the request as ASSEMBLED — messages, options
and the structured-output outcome together — and the caller emits both
`provider_request` and `structured_output_transport` from it. The event
that describes a request is now produced by the code that finished
building it, rather than by a caller guessing beforehand.

`callProviderWithIdleTimeout` moves to execution-provider-call.ts: the
file had grown two responsibilities — assembling a request and making
one safely — and crossed the 300-line ceiling. Re-exported from its old
home so existing importers keep one path.

Red-proved: reporting `conversationMessages` instead of the assembled
list fails the new test.
